requirements.txt
Pillow
moderngl
numpy
fastapi
uvicorn
python-multipart
requests
aiofiles